% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.R
\name{DR_statistic}
\alias{DR_statistic}
\title{DR statistic (R version)}
\usage{
DR_statistic(x)
}
\arguments{
\item{x}{an ape::phylo object}
}
\value{
vector of rates for each tip
}
\description{
Computes the Jetz DR rate. Code modified from fisse
}
\examples{
DR_statistic(ape::rcoal(10))
}
