% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DR_statistic_C}
\alias{DR_statistic_C}
\title{DR statistic (Rcpp version)}
\usage{
DR_statistic_C(x)
}
\arguments{
\item{x}{an ape::phylo object}
}
\value{
vector of rates for each tip
}
\description{
Computes the Jetz DR rate
}
\examples{
tree <- ape::rcoal(1000)
DR_statistic_C(tree)
}
