% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interval_node_statistic}
\alias{interval_node_statistic}
\title{Interval node statistic, with sampling times}
\usage{
interval_node_statistic(x, sample_times)
}
\arguments{
\item{x}{an ape::phylo object}

\item{sample_times}{a vector of ages}
}
\value{
matrix of rates for each tip, by sample time
}
\description{
Computes the interval node statistic, sampling at certain times in the psat
}
